All validations

+++ src/main/java/com/mrcruz/bikestore/service/ProductService.java
@@ -33,7 +33,7 @@ public class ProductService {
 		return productRepository.save(product);
 	}
 	
-	public void deleteProduct(Integer id) {
+	public void deleteProduct(Long id) {
 		productRepository.deleteById(id);
 	}
